refactor: refactoring code
