+++ src/hooks/useGtagEffect.ts
@@ -1,6 +1,6 @@
 import { useEffect } from 'react'
 import { useRouter } from 'next/router'
-import * as gtag from '@libs/gtag'
+import * as gtag from '@/src/libs/plugin/gtag'
 import CONFIG from '@/morethan-log.config'
 
 const useGtagEffect = () => {
