+++ src/libs/plugin/gtag.ts
