+++ src/libs/utils/index.ts
@@ -1,4 +1,4 @@
-export default function formatDate(date: any, local: any) {
+export function formatDate(date: any, local: any) {
   const d = new Date(date)
   const options: any = { year: 'numeric', month: 'short', day: 'numeric' }
   const res = d.toLocaleDateString(local, options)
