+++ src/pages/[slug].tsx
@@ -3,7 +3,7 @@ import { getAllPosts, getPostBlocks } from '@libs/notion'
 import Layout from '@components/Layout'
 import CONFIG from '../../morethan-log.config'
 import { NextPageWithLayout } from './_app'
-import { TPost } from '../types/post'
+import { TPost } from '../types'
 
 export async function getStaticPaths() {
   const posts = await getAllPosts({ includePages: true })
