+++ tsconfig.json
@@ -29,6 +29,6 @@
       "@custeomTypes/*": ["src/types/*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "src/libs/gtag.ts"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "src/libs/plugin/gtag.ts"],
   "exclude": ["node_modules"]
 }
